Moved DB calls to DataBase Folder
