+++ routes/administrador.js
@@ -1,10 +1,9 @@
 var express = require('express');
 var router = express.Router();
-const { connection } = require('../db');
-let Request = require('tedious').Request;
-const { TYPES } = require('tedious');
 
+//Estas son las llamadas a la base de datos que estamos utilizando en estas paginas de Administrador
 const { getListaPartidos } = require('../DataBase/getPartidos.js')
+
 
 router.get('/', function (req, res, next) {
     res.render('Admin/homepage');
